Display zoom percent rounded to nearest second decimal place.
This rounding is performed in the snprintf command. I believe C99 should
handle this by rounding to nearest. We do this to handle the edge case
where a value rounds to the same integer as actual size 100%

+++ src/image.h
@@ -191,17 +191,17 @@ void nqiv_image_manager_zoom_in_more(nqiv_image_manager* manager);
 void nqiv_image_manager_zoom_out_more(nqiv_image_manager* manager);
 
 /* Actually retrieve zoom rect. */
-void nqiv_image_manager_retrieve_zoomrect(nqiv_image_manager* manager,
-                                          const bool          do_zoom,
-                                          const bool          do_stretch,
-                                          SDL_Rect*           srcrect,
-                                          SDL_Rect*           dstrect);
+void   nqiv_image_manager_retrieve_zoomrect(nqiv_image_manager* manager,
+                                            const bool          do_zoom,
+                                            const bool          do_stretch,
+                                            SDL_Rect*           srcrect,
+                                            SDL_Rect*           dstrect);
 /* Calculate rations, actual size, fit amounts, etc' based on display and image
  * dimensions. */
-void nqiv_image_manager_calculate_zoom_parameters(nqiv_image_manager* manager,
-                                                  const SDL_Rect*     srcrect,
-                                                  const SDL_Rect*     dstrect);
-int  nqiv_image_manager_get_zoom_percent(nqiv_image_manager* manager);
+void   nqiv_image_manager_calculate_zoom_parameters(nqiv_image_manager* manager,
+                                                    const SDL_Rect*     srcrect,
+                                                    const SDL_Rect*     dstrect);
+double nqiv_image_manager_get_zoom_percent(nqiv_image_manager* manager);
 
 /* Thumbnail files are created based on the size displayed on screen, their
  * sizes determined by the Free Desktop Thumbnail Managing standard 0.9.0, which
